+++ src/entities/StudentEntity.ts
@@ -14,5 +14,7 @@ export default class StudentEntity implements Student {
     email = "";
     birthDate = "";
     phoneNumber = "";
+    enrollmentIdSecret = "";
+    isActive = true;
     constructor() {}
 }