+++ src/use/router/index.ts
@@ -10,11 +10,12 @@ import PageNotFound from "@/views/errors/404.vue";
 import { createRouter, createWebHistory } from "vue-router";
 const LoginView = () => import("@/views/common/Login.vue");
 const StudentCourseView = () => import("@/views/student/StudentCourseList.vue");
-const AllCourseView = () => import("@/views/shared/CourseList.vue");
+const AllCourseView = () => import("@/views/common/CourseList.vue");
 const AdminAccountListView = () => import("@/views/admin/AdminAccountList.vue");
-const CourseFormSuspenseWrapper = () => import("@/views/shared/EditCreateCourseForm.vue");
-const AccountFormSuspenseWrapper = () => import("@/views/admin/EditCreateAccountForm.vue");
-const ProfileWrapper = () => import("@/components/profile/Wrapper.vue");
+const CourseForm = () => import("@/views/shared/EditCreateCourseForm.vue");
+const AccountForm = () => import("@/views/admin/EditCreateAccountForm.vue");
+const PrivateProfile = () => import("@/views/common/PrivateProfile.vue");
+const PublicProfile = () => import("@/views/common/PublicProfile.vue");
 const Settings = () => import("@/views/common/Settings.vue");
 const StudentImmatricultaion = () => import("@/components/study/Immatriculation.vue");
 
@@ -39,6 +40,15 @@ const router = createRouter({
             component: AboutPage,
             meta: {
                 title: "About" + suffix,
+            },
+        },
+        {
+            path: "/course-catalog",
+            name: "courseCatalog",
+            props: { showAllCourses: true },
+            component: AllCourseView,
+            meta: {
+                title: "Course Catalog" + suffix,
             },
         },
         {
@@ -106,7 +116,7 @@ const router = createRouter({
             props: {
                 editMode: false,
             },
-            component: CourseFormSuspenseWrapper,
+            component: CourseForm,
             meta: {
                 title: "Course Creation" + suffix,
                 roles: ["Lecturer", "Admin"],
@@ -118,7 +128,7 @@ const router = createRouter({
             props: {
                 editMode: true,
             },
-            component: CourseFormSuspenseWrapper,
+            component: CourseForm,
             meta: {
                 title: "Course Editing" + suffix,
                 roles: ["Lecturer", "Admin"],
@@ -127,8 +137,7 @@ const router = createRouter({
         {
             path: "/user/:username",
             name: "profile.public",
-            props: { isPrivate: false },
-            component: ProfileWrapper,
+            component: PublicProfile,
             // The page title is set within the component depending on the username
             meta: {
                 roles: ["Admin", "Lecturer", "Student"],
@@ -137,8 +146,7 @@ const router = createRouter({
         {
             path: "/profile",
             name: "profile.private",
-            props: { isPrivate: true },
-            component: ProfileWrapper,
+            component: PrivateProfile,
             meta: {
                 title: "My Profile" + suffix,
                 roles: ["Admin", "Lecturer", "Student"],
@@ -158,7 +166,7 @@ const router = createRouter({
             props: {
                 editMode: false,
             },
-            component: AccountFormSuspenseWrapper,
+            component: AccountForm,
             meta: {
                 title: "Account Creation" + suffix,
                 roles: ["Admin"],
@@ -170,7 +178,7 @@ const router = createRouter({
             props: {
                 editMode: true,
             },
-            component: AccountFormSuspenseWrapper,
+            component: AccountForm,
             meta: {
                 title: "Account Editing" + suffix,
                 roles: ["Admin"],
@@ -247,7 +255,7 @@ router.beforeEach(async (to, from, next) => {
 
 router.afterEach(async (to, from, next) => {
     const store = useStore();
-    store.commit(MutationTypes.FORCE_CLOSE_BURGER_MENU, true);
+    store.commit(MutationTypes.FORCE_CLOSE_BURGER_MENU);
 });
 
 export default router;